moved toolbar view out of root folder
