+++ autosportlabs/racecapture/views/toolbar/toolbarview.py
